# cache type sizes in type-size limit visitor