@rust-log-analyzer 2024-07-03
The job **`x86_64-gnu-tools`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/9783934073/job/27013688851) [(plain)](https://github.com/rust-lang/rust/commit/96312fb87ccdfab9ebdf771e16c53890f9eeb679/checks/27013688851/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
   Compiling unicode-width v0.1.13
   Compiling getopts v0.2.21
   Compiling test v0.0.0 (/checkout/library/test)
   Compiling sysroot v0.0.0 (/checkout/library/sysroot)
error: reached the type-length limit while instantiating `<std::vec::IntoIter<u16> as Iterator>::try_fold::<(), {closure@<Enumerate<...> as Iterator>::try_fold::enumerate<'_, ..., ..., ..., ...>::{closure#0}}, ...>`
   |
   |
93 |         self.iter.try_fold(init, enumerate(&mut self.count, fold))
   |
   |
   = help: consider adding a `#![type_length_limit="59841203"]` attribute to your crate
   = note: the full type name has been written to '/checkout/obj/build/x86_64-unknown-linux-gnu/stage1-std/x86_64-unknown-linux-gnu/release/deps/test-13d0453d9d894b07.long-type.txt'
error: could not compile `test` (lib) due to 1 previous error
warning: build failed, waiting for other jobs to finish...
Build completed unsuccessfully in 0:04:56
+ set -e
+ set -e
+ cat /tmp/toolstate/toolstates.json
{"lld-wrapper":"test-fail"}
+ python3 ../x.py test --stage 2 check-tools
##[group]Building bootstrap
    Finished `dev` profile [unoptimized] target(s) in 0.04s
##[endgroup]
ERROR: Tool `book` was not recorded in tool state.
ERROR: Tool `nomicon` was not recorded in tool state.
ERROR: Tool `reference` was not recorded in tool state.
ERROR: Tool `rust-by-example` was not recorded in tool state.
ERROR: Tool `edition-guide` was not recorded in tool state.
ERROR: Tool `embedded-book` was not recorded in tool state.
  local time: Wed Jul  3 19:53:38 UTC 2024
  network time: Wed, 03 Jul 2024 19:53:38 GMT
##[error]Process completed with exit code 1.
Post job cleanup.
```

</details>