@rust-log-analyzer 2024-07-03
The job **`dist-x86_64-linux`** failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/actions/runs/9783955335/job/27013759061) [(plain)](https://github.com/rust-lang-ci/rust/commit/9ac628f2456372e15639a4ce65db7083e9a6df16/checks/27013759061/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
file:.git/config remote.origin.url=https://github.com/rust-lang-ci/rust
file:.git/config remote.origin.fetch=+refs/heads/*:refs/remotes/origin/*
file:.git/config gc.auto=0
file:.git/config http.https://github.com/.extraheader=AUTHORIZATION: basic ***
file:.git/config branch.try.remote=origin
file:.git/config branch.try.merge=refs/heads/try
file:.git/config submodule.library/backtrace.url=https://github.com/rust-lang/backtrace-rs.git
file:.git/config submodule.library/stdarch.active=true
file:.git/config submodule.library/stdarch.url=https://github.com/rust-lang/stdarch.git
file:.git/config submodule.src/doc/book.active=true
---
   Compiling getopts v0.2.21
[RUSTC-TIMING] unicode_width test:false 0.097
   Compiling test v0.0.0 (/checkout/library/test)
   Compiling sysroot v0.0.0 (/checkout/library/sysroot)
error: reached the type-length limit while instantiating `<std::vec::IntoIter<u16> as Iterator>::try_fold::<(), {closure@<Enumerate<...> as Iterator>::try_fold::enumerate<'_, ..., ..., ..., ...>::{closure#0}}, ...>`
   |
   |
93 |         self.iter.try_fold(init, enumerate(&mut self.count, fold))
   |
   |
   = help: consider adding a `#![type_length_limit="59841203"]` attribute to your crate
   = note: the full type name has been written to '/checkout/obj/build/x86_64-unknown-linux-gnu/stage1-std/x86_64-unknown-linux-gnu/release/deps/test-813daadf3317bef8.long-type.txt'
[RUSTC-TIMING] sysroot test:false 0.021
[RUSTC-TIMING] test test:false 0.694
error: could not compile `test` (lib) due to 1 previous error
warning: build failed, waiting for other jobs to finish...
---
Caused by:
    Command RUST_BACKTRACE=full python3 /checkout/x.py build --target x86_64-unknown-linux-gnu --host x86_64-unknown-linux-gnu --stage 2 library/std --rust-profile-generate /tmp/tmp-multistage/opt-artifacts/rustc-pgo --set llvm.thin-lto=false --set llvm.link-shared=true [at /checkout/obj] has failed with exit code Some(1)

Stack backtrace:
   0: <anyhow::Error>::msg::<alloc::string::String>
             at /rust/deps/anyhow-1.0.86/src/backtrace.rs:27:14
   1: <opt_dist::exec::CmdBuilder>::run
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/exec.rs:78:17
   2: <opt_dist::exec::Bootstrap>::run
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/exec.rs:179:9
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/main.rs:225:13
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/main.rs:225:13
   4: <opt_dist::timer::TimerSection>::section::<opt_dist::execute_pipeline::{closure#1}::{closure#0}, ()>
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/timer.rs:111:22
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/main.rs:214:9
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/main.rs:214:9
   6: <opt_dist::timer::TimerSection>::section::<opt_dist::execute_pipeline::{closure#1}, opt_dist::training::RustcPGOProfile>
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/timer.rs:111:22
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/main.rs:211:29
   8: opt_dist::main
             at /rustc/9ac628f2456372e15639a4ce65db7083e9a6df16/src/tools/opt-dist/src/main.rs:401:18
   9: <fn() -> core::result::Result<(), anyhow::Error> as core::ops::function::FnOnce<()>>::call_once
   9: <fn() -> core::result::Result<(), anyhow::Error> as core::ops::function::FnOnce<()>>::call_once
             at /rustc/75ac3b6331873133c4f7a10f2252afd6f3906c6a/library/core/src/ops/function.rs:250:5
  10: std::sys_common::backtrace::__rust_begin_short_backtrace::<fn() -> core::result::Result<(), anyhow::Error>, core::result::Result<(), anyhow::Error>>
             at /rustc/75ac3b6331873133c4f7a10f2252afd6f3906c6a/library/std/src/sys_common/backtrace.rs:155:18
  11: std::rt::lang_start::<core::result::Result<(), anyhow::Error>>::{closure#0}
             at /rustc/75ac3b6331873133c4f7a10f2252afd6f3906c6a/library/std/src/rt.rs:159:18
  12: core::ops::function::impls::<impl core::ops::function::FnOnce<A> for &F>::call_once
  13: std::panicking::try::do_call
             at /rustc/75ac3b6331873133c4f7a10f2252afd6f3906c6a/library/std/src/panicking.rs:559:40
  14: std::panicking::try
             at /rustc/75ac3b6331873133c4f7a10f2252afd6f3906c6a/library/std/src/panicking.rs:523:19
```

</details>