@lqd 2024-07-04
This should be ready to review, as at least a first step to reduce the problem.

@rustbot ready