@compiler-errors 2024-07-04
comparison url using the start of #125507's perf run and the end is this commit. Looks like we fixed everything unless I just can't read?

https://perf.rust-lang.org/compare.html?start=2db4ff40af2b9f93b6240dbd67ed7f2f34b19776&end=83217ae97e94eb43ce31396b49a938608f7f9c06&stat=instructions:u