@Kobzol 2024-07-09
Resolved a perf. regression from https://github.com/rust-lang/rust/pull/125507.

@rustbot label: +perf-regression-triaged